Change i++ to i+=1
