+++ src/index.html
@@ -6,11 +6,18 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" href="styles.css">
   <title>Space Shooter</title>
-    <script src="js/phaser.js"></script> 
   </head>
 
   <body>
+    <script src="js/phaser.js"></script> 
     <script src="js/Entities.js"></script>
+    <script src="js/playerClass.js"></script>
+    <script src="js/playerLaserClass.js"></script>
+    <script src="js/gunship.js"></script>
+    <script src="js/carriership.js"></script>
+    <script src="js/chasership.js"></script>
+    <script src="js/enemyLaser.js"></script>
+    <script src="js/scrollingBackground.js"></script>
     <script src="js/SceneMainMenu.js"></script>
     <script src="js/SceneMain.js"></script>
     <script src="js/SceneGameOver.js"></script>
