--- src/index.js
@@ -1,11 +1,0 @@
-import 'phaser';
-
-import { SimpleScene } from './scenes/simple-scene';
-
-const gameConfig = {
-  width: 680,
-  height: 400,
-  scene: SimpleScene,
-};
-
-new Phaser.Game(gameConfig);
