+++ src/js/SceneMainMenu.js
@@ -68,7 +68,7 @@ class SceneMainMenu extends Phaser.Scene {
     this.title1.setOrigin(0.5);
 
     this.backgrounds = [];
-    for (let i = 0; i < 5; i++) {
+    for (let i = 0; i < 5; i+=1) {
       const keys = ['sprBg0', 'sprBg1'];
       const key = keys[Phaser.Math.Between(0, keys.length - 1)];
       const bg = new ScrollingBackground(this, key, i * 10);
@@ -77,7 +77,7 @@ class SceneMainMenu extends Phaser.Scene {
   }
 
   update() {
-    for (let i = 0; i < this.backgrounds.length; i++) {
+    for (let i = 0; i < this.backgrounds.length; i+=1) {
       this.backgrounds[i].update();
     }
   }
