+++ src/js/carriership.js
@@ -1,0 +1,7 @@
+class CarrierShip extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, 'sprEnemy2', 'CarrierShip');
+    this.body.velocity.y = Phaser.Math.Between(50, 100);
+    this.play('sprEnemy2');
+  }
+}
