+++ src/js/enemyLaser.js
@@ -1,0 +1,6 @@
+class EnemyLaser extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, 'sprLaserEnemy0');
+    this.body.velocity.y = 200;
+  }
+}
